+++ controlplane/src/core/bufservices/PlatformService.ts
@@ -107,6 +107,7 @@ import {
   MoveGraphResponse,
   GenerateRouterTokenResponse,
   UpdateAISettingsResponse,
+  CreateBillingPortalSessionResponse,
 } from '@wundergraph/cosmo-connect/dist/platform/v1/platform_pb';
 import { isValidUrl } from '@wundergraph/cosmo-shared';
 import { DocumentNode, buildASTSchema, parse } from 'graphql';
@@ -4656,7 +4657,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           auditAction: 'subgraph_member.created',
           action: 'created',
           actorId: authContext.userId,
-          auditableType: 'subgraph',
+          auditableType: 'user',
           auditableDisplayName: user.email,
           actorDisplayName: authContext.userDisplayName,
           targetDisplayName: subgraph.name,
@@ -4711,6 +4712,19 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           authContext,
         });
 
+        const member = (await subgraphRepo.getSubgraphMembers(subgraph.id)).find(
+          (sm) => sm.subgraphMemberId === req.subgraphMemberId,
+        );
+
+        if (!member) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR_NOT_FOUND,
+              details: `The person is already not a member of the subgraph`,
+            },
+          };
+        }
+
         await subgraphRepo.removeSubgraphMember({ subgraphId: subgraph.id, subgraphMemberId: req.subgraphMemberId });
 
         await auditLogRepo.addAuditLog({
@@ -4718,8 +4732,8 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           auditAction: 'subgraph_member.deleted',
           action: 'deleted',
           actorId: authContext.userId,
-          auditableType: 'subgraph',
-          auditableDisplayName: subgraph.name,
+          auditableType: 'user',
+          auditableDisplayName: member.email,
           actorDisplayName: authContext.userDisplayName,
           targetDisplayName: subgraph.name,
           targetId: subgraph.id,
@@ -5234,6 +5248,8 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         const subgraphRepo = new SubgraphRepository(opts.db, authContext.organizationId);
         const orgRepo = new OrganizationRepository(opts.db, opts.billingDefaultPlanId);
 
+        req.namespace = req.namespace || DefaultNamespace;
+
         const federatedGraph = await fedgraphRepo.byName(req.name, req.namespace);
         if (!federatedGraph) {
           return {
@@ -5241,8 +5257,8 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
               code: EnumStatusCode.ERR_NOT_FOUND,
             },
             checks: [],
-            checksCountBasedOnDateRange: '0',
-            totalChecksCount: '0',
+            checksCountBasedOnDateRange: 0,
+            totalChecksCount: 0,
           };
         }
 
@@ -5266,8 +5282,21 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
               details: 'Invalid date range',
             },
             checks: [],
-            checksCountBasedOnDateRange: '0',
-            totalChecksCount: '0',
+            checksCountBasedOnDateRange: 0,
+            totalChecksCount: 0,
+          };
+        }
+
+        // check that the limit is less than the max option provided in the ui
+        if (req.limit > 50) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR,
+              details: 'Invalid limit',
+            },
+            checks: [],
+            checksCountBasedOnDateRange: 0,
+            totalChecksCount: 0,
           };
         }
 
@@ -5285,8 +5314,8 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
             code: EnumStatusCode.OK,
           },
           checks: checksData.checks,
-          checksCountBasedOnDateRange: checksData.checksCount.toString(),
-          totalChecksCount: totalChecksCount.toString(),
+          checksCountBasedOnDateRange: checksData.checksCount,
+          totalChecksCount,
         };
       });
     },
@@ -6334,6 +6363,8 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         const orgRepo = new OrganizationRepository(opts.db, opts.billingDefaultPlanId);
         const graphCompositionRepository = new GraphCompositionRepository(opts.db);
 
+        req.namespace = req.namespace || DefaultNamespace;
+
         const federatedGraph = await fedRepo.byName(req.fedGraphName, req.namespace);
 
         if (!federatedGraph) {
@@ -6343,6 +6374,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
               details: `Federated graph '${req.fedGraphName}' does not exist`,
             },
             compositions: [],
+            count: 0,
           };
         }
 
@@ -6366,6 +6398,19 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
               details: 'Invalid date range',
             },
             compositions: [],
+            count: 0,
+          };
+        }
+
+        // check that the limit is less than the max option provided in the ui
+        if (req.limit > 50) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR,
+              details: 'Invalid limit',
+            },
+            compositions: [],
+            count: 0,
           };
         }
 
@@ -6380,11 +6425,20 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           },
         });
 
+        const compositionsCount = await graphCompositionRepository.getGraphCompositionsCount({
+          fedGraphTargetId: federatedGraph.targetId,
+          dateRange: {
+            start: dateRange.start,
+            end: dateRange.end,
+          },
+        });
+
         return {
           response: {
             code: EnumStatusCode.OK,
           },
           compositions,
+          count: compositionsCount,
         };
       });
     },
@@ -6743,23 +6797,25 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         const billingRepo = new BillingRepository(opts.db);
         const plans = await billingRepo.listPlans();
 
-        if (!plans?.length) {
-          return {
-            response: {
-              code: EnumStatusCode.ERR,
-              details: 'No billing plans configured. Please contact support.',
-            },
-          };
-        }
+        if (opts.stripeSecretKey) {
+          if (!plans?.length) {
+            return {
+              response: {
+                code: EnumStatusCode.ERR,
+                details: 'No billing plans configured. Please contact support.',
+              },
+            };
+          }
 
-        // Validate the plan
-        if (plans?.length && !plans.some((plan) => plan.id === req.plan && 'stripePriceId' in plan)) {
-          return {
-            response: {
-              code: EnumStatusCode.ERR,
-              details: 'Invalid plan. Please contact support.',
-            },
-          };
+          // Validate the plan
+          if (plans?.length && !plans.some((plan) => plan.id === req.plan && 'stripePriceId' in plan)) {
+            return {
+              response: {
+                code: EnumStatusCode.ERR,
+                details: 'Invalid plan. Please contact support.',
+              },
+            };
+          }
         }
 
         await opts.keycloakClient.authenticateClient();
@@ -6808,11 +6864,15 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
               roles: ['admin'],
             });
 
-            const session = await billingService.createCheckoutSession({
-              organizationId: organization.id,
-              organizationSlug: organization.slug,
-              plan: req.plan,
-            });
+            let sessionId: string | undefined;
+            if (opts.stripeSecretKey) {
+              const session = await billingService.createCheckoutSession({
+                organizationId: organization.id,
+                organizationSlug: organization.slug,
+                plan: req.plan,
+              });
+              sessionId = session.id;
+            }
 
             const namespaceRepo = new NamespaceRepository(tx, organization.id);
             const ns = await namespaceRepo.create({
@@ -6837,7 +6897,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
 
             return {
               organization,
-              sessionId: session.id,
+              sessionId,
             };
           });
 
@@ -6885,6 +6945,16 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         const billingRepo = new BillingRepository(opts.db);
         const billingService = new BillingService(opts.db, billingRepo);
 
+        if (!opts.stripeSecretKey) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR,
+              details: 'Billing is not enabled. Please contact support.',
+            },
+            sessionId: '',
+          };
+        }
+
         const session = await billingService.createCheckoutSession({
           organizationId: authContext.organizationId,
           organizationSlug: authContext.organizationSlug,
@@ -6911,6 +6981,15 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         const billingRepo = new BillingRepository(opts.db);
         const billingService = new BillingService(opts.db, billingRepo);
         const auditLogRepository = new AuditLogRepository(opts.db);
+
+        if (!opts.stripeSecretKey) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR,
+              details: 'Billing is not enabled. Please contact support.',
+            },
+          };
+        }
 
         const plan = await billingRepo.getPlanById(req.plan);
         if (!plan) {
@@ -6951,10 +7030,21 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         method: ctx.method.name,
       });
 
-      return handleError<PlainMessage<CreateCheckoutSessionResponse>>(logger, async () => {
+      return handleError<PlainMessage<CreateBillingPortalSessionResponse>>(logger, async () => {
         const authContext = await opts.authenticator.authenticate(ctx.requestHeader);
         const billingRepo = new BillingRepository(opts.db);
         const billingService = new BillingService(opts.db, billingRepo);
+
+        if (!opts.stripeSecretKey) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR,
+              details: 'Billing is not enabled. Please contact support.',
+            },
+            sessionId: '',
+            url: '',
+          };
+        }
 
         const session = await billingService.createBillingPortalSession({
           organizationId: authContext.organizationId,