+++ studio/src/pages/[organizationSlug]/audit-log.tsx
@@ -9,6 +9,7 @@ import { EmptyState } from "@/components/empty-state";
 import { getDashboardLayout } from "@/components/layout/dashboard-layout";
 import { Button } from "@/components/ui/button";
 import { Loader } from "@/components/ui/loader";
+import { Pagination } from "@/components/ui/pagination";
 import {
   Select,
   SelectContent,
@@ -60,18 +61,6 @@ const AuditLogPage: NextPageWithLayout = () => {
     queryKey: [router.asPath, "GetAuditLogs", {}],
   });
 
-  const applyNewParams = useCallback(
-    (newParams: Record<string, string>) => {
-      router.push({
-        query: {
-          ...router.query,
-          ...newParams,
-        },
-      });
-    },
-    [router],
-  );
-
   if (isLoading) return <Loader fullscreen />;
 
   if (data?.response?.code === EnumStatusCode.ERROR_NOT_AUTHORIZED) {
@@ -98,77 +87,7 @@ const AuditLogPage: NextPageWithLayout = () => {
   return (
     <div className="flex h-full flex-col gap-y-4">
       <AuditLogTable logs={data?.logs} />
-      <div className="mr-2 flex justify-end">
-        <div className="flex items-center space-x-2">
-          <p className="text-sm font-medium">Rows per page</p>
-          <Select
-            value={`${limit}`}
-            onValueChange={(value) => {
-              applyNewParams({ pageSize: value });
-            }}
-          >
-            <SelectTrigger className="h-8 w-[70px]">
-              <SelectValue placeholder={`${limit}`} />
-            </SelectTrigger>
-            <SelectContent side="top">
-              {[10, 20, 30, 40, 50].map((pageSize) => (
-                <SelectItem key={pageSize} value={`${pageSize}`}>
-                  {pageSize}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-        </div>
-        <div className="flex w-[100px] items-center justify-center text-sm font-medium">
-          Page {noOfPages === 0 ? "0" : pageNumber} of {noOfPages}
-        </div>
-        <div className="flex items-center space-x-2">
-          <Button
-            variant="outline"
-            className="hidden h-8 w-8 p-0 lg:flex"
-            onClick={() => {
-              applyNewParams({ page: "1" });
-            }}
-            disabled={pageNumber === 1}
-          >
-            <span className="sr-only">Go to first page</span>
-            <DoubleArrowLeftIcon className="h-4 w-4" />
-          </Button>
-          <Button
-            variant="outline"
-            className="h-8 w-8 p-0"
-            onClick={() => {
-              applyNewParams({ page: (pageNumber - 1).toString() });
-            }}
-            disabled={pageNumber === 1}
-          >
-            <span className="sr-only">Go to previous page</span>
-            <ChevronLeftIcon className="h-4 w-4" />
-          </Button>
-          <Button
-            variant="outline"
-            className="h-8 w-8 p-0"
-            onClick={() => {
-              applyNewParams({ page: (pageNumber + 1).toString() });
-            }}
-            disabled={pageNumber === noOfPages || noOfPages === 0}
-          >
-            <span className="sr-only">Go to next page</span>
-            <ChevronRightIcon className="h-4 w-4" />
-          </Button>
-          <Button
-            variant="outline"
-            className="hidden h-8 w-8 p-0 lg:flex"
-            onClick={() => {
-              applyNewParams({ page: noOfPages.toString() });
-            }}
-            disabled={pageNumber === noOfPages || noOfPages === 0}
-          >
-            <span className="sr-only">Go to last page</span>
-            <DoubleArrowRightIcon className="h-4 w-4" />
-          </Button>
-        </div>
-      </div>
+      <Pagination limit={limit} noOfPages={noOfPages} pageNumber={pageNumber} />
     </div>
   );
 };